added wake up before commands to prevent adbms go to sleep

+++ src/main.cpp
@@ -6,7 +6,7 @@
 #define N 1
 int8_t pins[4] = {14,12,13,15};//tab for custom SPI Pins. sck, miso, mosi, cs is a pin order. In this tab i use HSPI port SPI pins
 //ADBMS1818 adbms(pins); //constructor with modified pins
-ADBMS1818 adbms((uint16_t)HSPI, 15); //construcot with modified CS pin and spi port
+ADBMS1818 adbms((uint8_t)HSPI, (uint8_t)15); //construcot with modified CS pin and spi port
 //ADBMS1818 adbms(15); //constructor with default SPI and CS pin modifed
 void setup() {
   // put your setup code here, to run once:
